Put alt handling code in FormBase. Fixes #3584.
